requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
sqlmodel>=0.0.8
python-jose>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.0
pytest>=7.0.0
httpx>=0.23.0

# Pin bcrypt to 3.2.0 because passlib 1.7.4 expects bcrypt to expose
# the module attribute `__about__` for version detection. Newer bcrypt
# 4.x releases do not expose `__about__` and cause a non-fatal warning
# from passlib ("error reading bcrypt version"). Pinning ensures clean
# logs and compatibility with the installed passlib version.
bcrypt==3.2.0
